Upgrade

+++ cettia/two/client/src/app/edit/edit.page.ts
@@ -12,7 +12,7 @@ import {ITodo} from '../protos/changeevent';
 })
 export class EditPage implements OnInit {
 
-  todo: ITodo;
+  todo: ITodo | undefined;
 
   constructor(private readonly navCtrl: NavController,
               private readonly route: ActivatedRoute,
@@ -20,7 +20,7 @@ export class EditPage implements OnInit {
 
   }
 
-  ngOnInit() {
+  ngOnInit(): void {
     const id = this.route.snapshot.paramMap.get('id');
     if (id) {
       this.todo = this.todoService.getTodo(id);
@@ -33,13 +33,17 @@ export class EditPage implements OnInit {
     }
   }
 
-  save() {
-    this.todoService.save(this.todo);
+  save(): void {
+    if (this.todo) {
+      this.todoService.save(this.todo);
+    }
     this.navCtrl.navigateBack(['home']);
   }
 
-  deleteTodo() {
-    this.todoService.deleteTodo(this.todo);
+  deleteTodo(): void {
+    if (this.todo) {
+      this.todoService.deleteTodo(this.todo);
+    }
     this.navCtrl.navigateBack(['home']);
   }
